Poprawiony projekt - domain, repository, service, inMemoryRepository

+++ Cookbook/src/main/java/com/ania/cookbook/infrastructure/repositories/InMemoryRecipeRepository.java
@@ -1,62 +1,105 @@
 package com.ania.cookbook.infrastructure.repositories;
 
+import com.ania.cookbook.domain.exceptions.RecipeNotFoundException;
+import com.ania.cookbook.domain.exceptions.RecipeValidationException;
+import com.ania.cookbook.domain.model.Category;
 import com.ania.cookbook.domain.repositories.recipe.DeleteRecipe;
 import com.ania.cookbook.domain.repositories.recipe.ReadRecipe;
 import com.ania.cookbook.domain.repositories.recipe.SaveRecipe;
 import com.ania.cookbook.domain.repositories.recipe.UpdateRecipe;
-import com.ania.cookbook.infrastructure.persistence.entity.ProductEntity;
-import com.ania.cookbook.infrastructure.persistence.entity.RecipeEntity;
+import com.ania.cookbook.domain.model.Recipe;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
+
 import java.time.Instant;
 import java.util.HashMap;
 import java.util.Optional;
 import java.util.UUID;
+
+@RequiredArgsConstructor
 @Component
 public class InMemoryRecipeRepository implements SaveRecipe, ReadRecipe, UpdateRecipe, DeleteRecipe {
-    private final HashMap<UUID, RecipeEntity> inMemoryRepository = new HashMap<>();
-    private final HashMap<String, RecipeEntity> inMemoryRepositoryByName = new HashMap<>();
+    private final HashMap<UUID, Recipe> inMemoryRepository = new HashMap<>();
+
     @Override
-    public RecipeEntity saveRecipe(RecipeEntity recipe){
+    public Recipe saveRecipe(Recipe recipe){
+        if (recipe == null) {
+            throw new RecipeValidationException("Recipe cannot be null");}
+        if (inMemoryRepository.containsKey(recipe.getRecipeId())) {
+            throw new RecipeValidationException("A recipe with the given Id already exists.");
+        }
         inMemoryRepository.put(recipe.getRecipeId(), recipe);
-        inMemoryRepositoryByName.put(recipe.getRecipeName(), recipe);
         return recipe;
     }
+
     @Override
-    public Optional<RecipeEntity> findRecipeById(UUID id){
+    public Optional<Recipe> findRecipeById(UUID id){
         return Optional.ofNullable(inMemoryRepository.get(id));
     }
+
     @Override
-    public Optional<RecipeEntity> findRecipeByName(String name){
-        return Optional.ofNullable(inMemoryRepositoryByName.get(name));
+    public boolean existsRecipeById(UUID recipeId) {
+        return inMemoryRepository.containsKey(recipeId);
+    }
+
+    @Override
+    public Optional<Recipe> findRecipeByName(String name){
+        return inMemoryRepository.values().stream()
+                .filter(recipe -> recipe.getRecipeName().equalsIgnoreCase(name))
+                .findFirst();
+    }
+
+    @Override
+    public boolean existsRecipeByName(String recipeName) {
+        return inMemoryRepository.values().stream()
+                .anyMatch(recipe -> recipe.getRecipeName().equalsIgnoreCase(recipeName));
+    }
+
+    @Override
+    public Optional<Recipe> findRecipeByCategory(Category category){
+        return inMemoryRepository.values().stream()
+                .filter(recipe -> recipe.getCategory() == category)
+                .findFirst();
     }
     @Override
-    public Optional<RecipeEntity> findRecipeByCategoryContains(String category){
-        return inMemoryRepository.values().stream().filter(recipe -> recipe.getCategories().stream()
-                .anyMatch(cat ->cat.toString().equalsIgnoreCase(category))).findFirst();
-    }
-    @Override
-    public Optional<RecipeEntity> findRecipeByIngredientContains(String ingredient){
+    public Optional<Recipe> findRecipeByIngredientContains(String ingredient){
         return inMemoryRepository.values().stream().filter(recipe -> recipe.getIngredients()
-                .stream().anyMatch(ing -> ing.getProductEntity().map(ProductEntity::getProductName)
-                        .orElse("").equalsIgnoreCase(ingredient))).findFirst();
+                .stream().anyMatch(ing -> ing.getProduct().getProductName()
+                        .equalsIgnoreCase(ingredient))).findFirst();
     }
-    @Override
-    public Optional<RecipeEntity> findRecipeByProduct(ProductEntity product){
-        return Optional.ofNullable(inMemoryRepository.get(product.getProductId()));
+   @Override
+    public Optional<Recipe> findRecipeByProductId(UUID productId){
+        return inMemoryRepository.values().stream().filter(recipe -> recipe.getIngredients()
+                .stream().anyMatch(ing -> ing.getProduct().getProductId().equals(productId)))
+                .findFirst();
     }
+
     @Override
-    public Optional<RecipeEntity> findRecipeByCreatedAfter(Instant created){
+    public Optional<Recipe> findRecipeByProductName(String productName){
+        return inMemoryRepository.values().stream().filter(recipe -> recipe.getIngredients()
+                        .stream().anyMatch(ing -> ing.getProduct().getProductName().equalsIgnoreCase(productName)))
+                .findFirst();
+    }
+
+    @Override
+    public Optional<Recipe> findRecipeByCreatedAfter(Instant created){
         return inMemoryRepository.values().stream()
                 .filter(recipe ->recipe.getCreated().isAfter(created))
                 .findFirst();
     }
     @Override
-    public RecipeEntity updateRecipe(RecipeEntity recipe){
+    public Recipe updateRecipe(Recipe recipe){
+        if (!inMemoryRepository.containsKey(recipe.getRecipeId())) {
+            throw new RecipeNotFoundException("Recipe with Id does not exist.");
+        }
         inMemoryRepository.put(recipe.getRecipeId(), recipe);
         return recipe;
     }
     @Override
     public void deleteRecipeById(UUID id){
+        if (!inMemoryRepository.containsKey(id)) {
+            throw new RecipeNotFoundException("Recipe with Id does not exist.");
+        }
         inMemoryRepository.remove(id);
     }
 }
